+++ CHANGES.md
@@ -1,3 +1,11 @@
+### 20240426
+ * Improved cm run script --repro function to dump dependencies, versions and generate README
+
+### 20240425
+ * CM script automation: fixed dumping versions (variations didn't have _ prefix)
+
+---
+
 Since March 2023, all updates to CM automations are submitted via PRs.
 You can follow our PRs at
 * https://github.com/ctuning/mlcommons-ck/commits/master